Resolved Perfective Issue 1
Resolved perfective issue 1 by improving the readability of error messages in index.js, as well as assigning a timeout so the error message does not remain on screen for the remainder of the game. As a result, the error messages are more user-friendly, and can be interpreted by the average player without issue.